backend/requirements.txt
Django==2.2.19
djangorestframework==3.13.1
flake8==4.0.1
importlib-metadata==4.2.0
mccabe==0.6.1
Pillow==8.4.0
pkg_resources==0.0.0
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
zipp==3.6.0